sa without queue ready

+++ ulisa/bkp/sa_v5/pe.cpp
@@ -5,7 +5,7 @@ PE::PE():li(0),tw(0),ri(0),bw(0),ba(0){}
 void PE::compute() {
     ri = li;
     bw = tw;
-    ba += li * tw;
+    ba += static_cast<int16_t>(li) * static_cast<int16_t>(tw);
 }
 
 void PE::reset(){
